+++ docs/Whats-New.md
@@ -336,7 +336,7 @@ New:
 - Wall overlay `Palette` support (by Starkku)
 - Show designator & inhibitor range (by Morton)
 - Owner-only sound on unit creation (by Fryone)
-- `Secondary` will now be used against walls if `Primary` weapon Warhead has `Wall=false` (by Starkku)
+- Allow using `Secondary` weapon against walls if `Primary` cannot target them (by Starkku)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)